+++ hosts/zaneyos-oem/variables.nix
@@ -3,6 +3,12 @@
   gitUsername = "Don Williams";
   gitEmail = "don.e.williams@gmail.com";
 
+  # Set Displau Manager
+  # `tui` for Text login
+  # `sddm` for graphical GUI
+  # SDDM background is set with stylixImage
+  displayManager = "sddm";
+
   # Hyprland Settings
   extraMonitorSettings = "monitor=Virtual-1,1920x1080@75,auto,1";
 
@@ -25,14 +31,24 @@
   # Enable Printing Support
   printEnable = false;
 
+  # Enable Thunar GUI File Manager
+  thunarEnable = true;
+
   # Set Stylix Image
+  # This will set your color palette
+  # Default background
+  # Add new images to ~/zaneyos/wallpapers
   stylixImage = ../../wallpapers/mountainscapedark.jpg;
   #stylixImage = ../../wallpapers/AnimeGirlNightSky.jpg;
   #stylixImage = ../../wallpapers/Anime-Purple-eyes.png;
+  #stylixImage = ../../wallpapers/Rainnight.jpg;
+  #stylixImage = ../../wallpapers/zaney-wallpaper.jpg;
+  #stylixImage = ../../wallpapers/nix-wallpapers-strips-logo.jpg;
+  #stylixImage = ../../wallpapers/beautifulmountainscape.jpg;
 
   # Set Waybar
   # Available options:
-  #waybarChoice = ../../modules/home/waybar/Jerry-simple.nix;
+  #waybarChoice= ../../modules/home/waybar/Jerry-simple.nix;
   #waybarChoice = ../../modules/home/waybar/Jerry-curved.nix;
   #waybarChoice = ../../modules/home/waybar/Jerry-waybar.nix;
   #waybarChoice = ../../modules/home/waybar/waybar-ddubs.nix;
@@ -45,7 +61,7 @@
   # animations-dynamic.nix (ml4w project)
   # animations-moving.nix (ml4w project)
   animChoice = ../../modules/home/hyprland/animations-end4.nix;
-
-  # Enable Thunar GUI File Manager
-  thunarEnable = true;
+  #animChoice = ../../modules/home/hyprland/animations-def.nix;
+  #animChoice = ../../modules/home/hyprland/animations-dynamic.nix;
+  #animChoice = ../../modules/home/hyprland/animations-moving.nix;
 }