+++ modules/home/nvf.nix
@@ -2,9 +2,8 @@
   inputs,
   config,
   ...
-}:
-{
-  imports = [ inputs.nvf.homeManagerModules.default ];
+}: {
+  imports = [inputs.nvf.homeManagerModules.default];
 
   programs.nvf = {
     enable = true;
@@ -52,55 +51,55 @@
       keymaps = [
         {
           key = "jk";
-          mode = [ "i" ];
+          mode = ["i"];
           action = "<ESC>";
           desc = "Exit insert mode";
         }
         {
           key = "<leader>nh";
-          mode = [ "n" ];
+          mode = ["n"];
           action = ":nohl<CR>";
           desc = "Clear search highlights";
         }
         {
           key = "<leader>ff";
-          mode = [ "n" ];
+          mode = ["n"];
           action = "<cmd>Telescope find_files<cr>";
           desc = "Search files by name";
         }
         {
           key = "<leader>lg";
-          mode = [ "n" ];
+          mode = ["n"];
           action = "<cmd>Telescope live_grep<cr>";
           desc = "Search files by contents";
         }
         {
           key = "<leader>fe";
-          mode = [ "n" ];
+          mode = ["n"];
           action = "<cmd>Neotree toggle<cr>";
           desc = "File browser toggle";
         }
         {
           key = "<C-h>";
-          mode = [ "i" ];
+          mode = ["i"];
           action = "<Left>";
           desc = "Move left in insert mode";
         }
         {
           key = "<C-j>";
-          mode = [ "i" ];
+          mode = ["i"];
           action = "<Down>";
           desc = "Move down in insert mode";
         }
         {
           key = "<C-k>";
-          mode = [ "i" ];
+          mode = ["i"];
           action = "<Up>";
           desc = "Move up in insert mode";
         }
         {
           key = "<C-l>";
-          mode = [ "i" ];
+          mode = ["i"];
           action = "<Right>";
           desc = "Move right in insert mode";
         }
@@ -110,7 +109,7 @@
 
       spellcheck = {
         enable = true;
-        languages = [ "en" ];
+        languages = ["en"];
         programmingWordlist.enable = true;
       };
 
@@ -142,7 +141,10 @@
         };
         html.enable = true;
         lua.enable = true;
-        css.enable = false;
+        css = {
+          enable = true;
+          format.type = "prettierd";
+        };
         typst.enable = true;
         rust = {
           enable = true;